+++ tests/linux/Linux.sh
@@ -28,11 +28,14 @@ function buildPackage() {
     if [ "$DISTRO" = "debian-trixie" ]; then
       sudo docker build -t linux-debian-trixie -f src/linux/Dockerfile-debian-trixie .
     fi
-    if [ "$DISTRO" = "fedora-40" ]; then
-      sudo docker build -t linux-fedora-40 -f src/linux/Dockerfile-fedora-40 .
-    fi
     if [ "$DISTRO" = "fedora-41" ]; then
       sudo docker build -t linux-fedora-41 -f src/linux/Dockerfile-fedora-41 .
+    fi
+    if [ "$DISTRO" = "fedora-42" ]; then
+      sudo docker build -t linux-fedora-42 -f src/linux/Dockerfile-fedora-42 .
+    fi
+    if [ "$DISTRO" = "fedora-43" ]; then
+      sudo docker build -t linux-fedora-43 -f src/linux/Dockerfile-fedora-43 .
     fi
     if [ "$DISTRO" = "rhel-8" ]; then
       sudo docker build -t linux-rhel-8 -f src/linux/Dockerfile-rhel-8 .
@@ -62,11 +65,14 @@ function createContainer() {
     if [ "$DISTRO" = "debian-trixie" ]; then
       sudo docker run -v /tmp/debian-trixie:/data linux-debian-trixie
     fi
-    if [ "$DISTRO" = "fedora-40" ]; then
-      sudo docker run -v /tmp/fedora-40:/data linux-fedora-40
-    fi
     if [ "$DISTRO" = "fedora-41" ]; then
       sudo docker run -v /tmp/fedora-41:/data linux-fedora-41
+    fi
+    if [ "$DISTRO" = "fedora-42" ]; then
+      sudo docker run -v /tmp/fedora-42:/data linux-fedora-42
+    fi
+    if [ "$DISTRO" = "fedora-43" ]; then
+      sudo docker run -v /tmp/fedora-43:/data linux-fedora-43
     fi
     if [ "$DISTRO" = "rhel-8" ]; then
       sudo docker run -v /tmp/rhel-8:/data linux-rhel-8
@@ -86,7 +92,7 @@ function createContainer() {
 # Retrieve $DISTRO from the user
 
 function retrieveDistro() {
-  echo "Which distro do you want to use? (ubuntu, debian-bookworm, debian-trixie, fedora-40, fedora-41, rhel-8, rhel-9, rhel-10)"
+  echo "Which distro do you want to use? (ubuntu, debian-bookworm, debian-trixie, fedora-41, fedora-42, fedora-43, rhel-8, rhel-9, rhel-10)"
   read -r DISTRO
 }